Added some traits classes and variadic template functions to deal with iterating over the control types.
By doing this the main window has absolutely no need to know about the
internals of the controls. In particular it no longer needs to know the
names of the types.

If a new type is added then its type should be added to the list of
types in the CONTROLPOINTERTYPES. A few other things will be needed like
adding the apropriate traits class and ensureing that the control has
the ItemType typedef, but these should be pointed out by the compiler if
they are missing.